// src/app/model/station.ts
export class Station {
    id: number;
    name: string;
    users: any;
    data: any;
    expenses: any;
    wetproductsales: any;
    dryproductsales: any;
    wetproductstocks: any;
    dryproductstocks: any;
    wetproductpurchases: any;
    dryproductsupplies: any;
    date_sold: Date;
    main: any;
}