// src/app/model/employee.ts
export class Employee {
    id: number;
    employee_level_id: number;
    station_id: number;
    level: string;
    firstName: string;
    secondName: string;
    lastName: string;
    salary: number;
    phone: number;
    address: string;
    date_of_birth: Date;
    date_hired: Date;
    station: string;
    name: string;
}